.github/actions/detect-exposed-secrets/exposed_secrets_detection/requirements-dev.txt
pytest
pyyaml